split package scripts

+++ scripts/package-json-version.ts
@@ -1,10 +1,8 @@
-import { readFile, writeFile, readdir } from 'node:fs/promises';
-import { dirname, resolve } from 'node:path';
-import { fileURLToPath } from 'node:url';
+import { readFile, writeFile } from 'node:fs/promises';
+import { existsSync } from 'node:fs';
+import { resolve } from 'node:path';
 import type { PackageJson } from 'type-fest';
-
-const __dirname = dirname(fileURLToPath(import.meta.url));
-const rootDir = resolve(__dirname, '..');
+import { packageDir, rootDir } from './utils/config.js';
 
 console.log('Updating Vaadin web components versions in "package.json"...');
 
@@ -14,34 +12,32 @@ const workspacePackageJson = JSON.parse(await readFile(workspacePackageJsonPath,
 const componentsVersion = workspacePackageJson.version;
 console.log(`Using "${componentsVersion}" as the version for web components.`);
 
-const packagesDir = resolve(rootDir, 'packages');
-const packages = await readdir(packagesDir);
+const packageJsonPath = resolve(packageDir, 'package.json');
 
-for (const packageName of packages) {
-  const packageJsonPath = resolve(packagesDir, packageName, 'package.json');
-  const packageJson = JSON.parse(await readFile(packageJsonPath, 'utf8')) as PackageJson;
+const packageJson = JSON.parse(await readFile(packageJsonPath, 'utf8')) as PackageJson;
 
-  Object.keys(packageJson.dependencies!).forEach((name) => {
-    if (!name.startsWith('@vaadin/')) {
-      return;
-    }
+Object.keys(packageJson.dependencies!).forEach((name) => {
+  if (!name.startsWith('@vaadin/')) {
+    return;
+  }
 
-    packageJson.version = componentsVersion;
-    packageJson.dependencies![name] = componentsVersion;
-  });
+  packageJson.version = componentsVersion;
+  packageJson.dependencies![name] = componentsVersion;
+});
 
-  await writeFile(packageJsonPath, JSON.stringify(packageJson, null, 2), 'utf8');
-}
+await writeFile(packageJsonPath, JSON.stringify(packageJson, null, 2), 'utf8');
 
 // Update version metadata in the sources
 
-const versionSourceFile = 'packages/react-components/src/utils/createComponent.ts';
-console.log(`Updating version metadata in "${versionSourceFile}"...`);
-const versionSourcePath = resolve(rootDir, versionSourceFile);
-const versionSource = await readFile(versionSourcePath, 'utf8');
-const newVersionSource = versionSource.replace(
-  /version:.+,/g,
-  `version: /* updated-by-script */ '${componentsVersion}',`,
-);
+const versionSourceFile = resolve(packageDir, 'src/utils/createComponent.ts');
+if (existsSync(versionSourceFile)) {
+  console.log(`Updating version metadata in "${versionSourceFile}"...`);
+  const versionSourcePath = resolve(rootDir, versionSourceFile);
+  const versionSource = await readFile(versionSourcePath, 'utf8');
+  const newVersionSource = versionSource.replace(
+    /version:.+,/g,
+    `version: /* updated-by-script */ '${componentsVersion}',`,
+  );
 
-await writeFile(versionSourcePath, newVersionSource, 'utf8');
+  await writeFile(versionSourcePath, newVersionSource, 'utf8');
+}
